Yeet legacy events api stuff, except the most used part
