+++ apps/events/urls.py
@@ -48,14 +48,3 @@ urlpatterns = [
 # API v1
 router = SharedAPIRootRouter()
 router.register("events", views.EventViewSet, basename="events")
-router.register("attendees", views.AttendeeViewSet, basename="attendees")
-router.register(
-    prefix="registration/attendance-events",
-    viewset=views.RegistrationAttendaceEventViewSet,
-    basename="registration_attendance_events",
-)
-router.register(
-    prefix="registration/user-attendees",
-    viewset=views.RegistrationAttendeeViewSet,
-    basename="registration_attendees",
-)
